refactor: extract module-level path helpers → _path_utils.py
Move the ten module-level path helpers out of pdsfile.py into a new private
module: construct_category_list, logical_path_from_abspath, _clean_join,
_clean_abspath, _clean_glob, _needs_glob, repair_case, formatted_file_size,
abspath_for_logical_path and selected_path_from_path. Their bodies are
byte-for-byte identical to the ones they replace.

Two module constants move with them, as the free-variable sweep requires:
FILE_BYTE_UNITS (public, so pdsfile.py re-exports it) and _GLOB_CACHE_SIZE
(private, referenced only by _clean_glob's lru_cache decorator).
PATH_EXISTS_CACHE_SIZE stays; its consumer moves in PR-17.

pdsfile.py imports every moved name back, so pdsfile.pdsfile.<name> resolves
exactly as before and the API manifest is unchanged. glob and math are no
longer referenced there but are frozen members of that module's surface, so
they stay bound in the explicit re-export form.

_path_utils.py's ruff entry is a split of pdsfile.py's: E701 and F841 are
pre-existing violations of the moved lines and both codes remain on pdsfile.py
for the lines that stayed, so the ratchet gains no code.
